+++ src/XamEffects.iOS/Initialize.cs
@@ -1,12 +1,12 @@
 ﻿using XamEffects.iOS.GestureCollectors;
+using XamEffects.iOS.Renderers;
 
 namespace XamEffects.iOS {
     public static class Effects {
         public static void Init() {
             CommandsPlatform.Init();
             TouchEffectPlatform.Init();
-            LongTapGestureCollector.Init();
-            TapGestureCollector.Init();
+            BorderViewRenderer.Link();
         }
     }
 }