+++ src/XamEffects.iOS/TouchEffectPlatform.cs
@@ -8,6 +8,10 @@ using Xamarin.Forms.Platform.iOS;
 using XamEffects;
 using XamEffects.iOS;
 using XamEffects.iOS.GestureCollectors;
+using XamEffects.iOS.GestureRecognizers;
+using System;
+using System.Linq;
+using Foundation;
 
 [assembly: ExportEffect(typeof(TouchEffectPlatform), nameof(TouchEffect))]
 
@@ -17,22 +21,52 @@ namespace XamEffects.iOS {
         public UIView View => Control ?? Container;
 
         UIView _layer;
-        double _alpha;
-        CancellationTokenSource _cancellation;
+        nfloat _alpha;
 
         protected override void OnAttached() {
             View.UserInteractionEnabled = true;
+            _layer = new UIView {
+                UserInteractionEnabled = false,
+                Opaque = false,
+                Alpha = 0,
+                TranslatesAutoresizingMaskIntoConstraints = false
+            };
 
-            TapGestureCollector.Add(View, TapAction);
-            LongTapGestureCollector.Add(View, LongTapAction);
             UpdateEffectColor();
+            TouchGestureCollector.Add(View, OnTouch);
+
+            View.AddSubview(_layer);
+            View.BringSubviewToFront(_layer);
+            _layer.TopAnchor.ConstraintEqualTo(View.TopAnchor).Active = true;
+            _layer.LeftAnchor.ConstraintEqualTo(View.LeftAnchor).Active = true;
+            _layer.BottomAnchor.ConstraintEqualTo(View.BottomAnchor).Active = true;
+            _layer.RightAnchor.ConstraintEqualTo(View.RightAnchor).Active = true;
         }
 
         protected override void OnDetached() {
-            TapGestureCollector.Delete(View, TapAction);
-            LongTapGestureCollector.Delete(View, LongTapAction);
+            TouchGestureCollector.Delete(View, OnTouch);
+            _layer?.RemoveFromSuperview();
             _layer?.Dispose();
-            _layer = null;
+        }
+
+        void OnTouch(TouchGestureRecognizer.TouchArgs e) {
+            switch (e.State) {
+                case TouchGestureRecognizer.TouchState.Started:
+                    BringLayer();
+                    break;
+
+                case TouchGestureRecognizer.TouchState.Ended:
+                    EndAnimation();
+                    break;
+
+                case TouchGestureRecognizer.TouchState.Cancelled:
+                    if (!IsDisposed && _layer != null) {
+                        _layer.Layer.RemoveAllAnimations();
+                        _layer.Alpha = 0;
+                    }
+
+                    break;
+            }
         }
 
         protected override void OnElementPropertyChanged(PropertyChangedEventArgs e) {
@@ -43,55 +77,29 @@ namespace XamEffects.iOS {
             }
         }
 
-        async void LongTapAction(UIGestureRecognizerState state, bool inside) {
-            switch (state) {
-                case UIGestureRecognizerState.Began:
-                    await TapAnimation(0.3, 0, _alpha, false);
-                    break;
-                case UIGestureRecognizerState.Ended:
-                case UIGestureRecognizerState.Cancelled:
-                case UIGestureRecognizerState.Failed:
-                    await TapAnimation(0.3, _alpha);
-                    break;
-            }
-        }
-
-        async void TapAction() {
-            await TapAnimation(0.2, _alpha, 0);
-        }
-
         void UpdateEffectColor() {
-            _layer?.Dispose();
-            _layer = null;
-
             var color = TouchEffect.GetColor(Element);
             if (color == Color.Default) {
                 return;
             }
 
-            _alpha = color.A < 1.0 ? 1 : 0.3;
-            _layer = new UIView {BackgroundColor = color.ToUIColor(), UserInteractionEnabled = false,};
+            _alpha = color.A < 1.0 ? 1 : (nfloat)0.3;
+            _layer.BackgroundColor = color.ToUIColor();
         }
 
-        async Task TapAnimation(double duration, double start = 1, double end = 0, bool remove = true) {
+        void BringLayer() {
+            _layer.Layer.RemoveAllAnimations();
+            _layer.Alpha = _alpha;
+            View.BringSubviewToFront(_layer);
+        }
+
+        void EndAnimation() {
             if (!IsDisposed && _layer != null) {
-                _cancellation?.Cancel();
-                _cancellation = new CancellationTokenSource();
-
-                var token = _cancellation.Token;
-
-                _layer.Frame = new CGRect(0, 0, Container.Bounds.Width, Container.Bounds.Height);
-                Container.AddSubview(_layer);
-                Container.BringSubviewToFront(_layer);
-                _layer.Alpha = (float) start;
-                await UIView.AnimateAsync(duration,
-                    () => {
-                        if (!token.IsCancellationRequested && !IsDisposed)
-                            _layer.Alpha = (float) end;
-                    });
-                if (remove && !IsDisposed && !token.IsCancellationRequested) {
-                    _layer?.RemoveFromSuperview();
-                }
+                _layer.Layer.RemoveAllAnimations();
+                UIView.Animate(0.225,
+                () => {
+                    _layer.Alpha = 0;
+                });
             }
         }