+++ src/XamEffects/Initialize.cs
@@ -1,5 +1,7 @@
-﻿namespace XamEffects {
+﻿using System;
+namespace XamEffects {
     public static class Effects {
+        [Obsolete("Not needed with usual Linking")]
         public static void Init() {
             TouchEffect.Init();
             Commands.Init();